Creating a custom error class

+++ src/storage/group/groupCreate.ts
@@ -1,13 +1,22 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
+
 import { GROUP_COLLECTION } from "../storageConfig";
 import { groupsGetAll } from "./groupsGetAll";
+import { AppError } from "@utils/AppError";
 
 export async function groupCreate(newGroup: string) {
   try {
     const storedGroups = await groupsGetAll();
 
+    const groupAllReadyExists = storedGroups.includes(newGroup);
+
+    if(groupAllReadyExists) {
+      throw new AppError('Já existe um grupo cadastrado com esse nome.');
+    }
+
     const storage = JSON.stringify([...storedGroups, newGroup])
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
+
   } catch (error) {
     throw error;
   }
